refactor: query, mutation 폴더구조 정리
