+++ packages/react-native/src/apis/queries/common/useProfileQuery.ts
@@ -2,7 +2,7 @@ import { useRef } from 'react';
 import { useSuspenseQuery } from '@tanstack/react-query';
 import tinycolor from 'tinycolor2';
 import { AppStorage } from '@/utils/storage';
-import useAuthAxios from '../useAuthAxios';
+import useAuthAxios from '../../useAuthAxios';
 import QUERY_KEYS from '@/constants/QUERY_KEYS';
 import { ServerResponse } from '@/types/response';
